% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_simply_logo}
\alias{add_simply_logo}
\title{A function for creating simplystats themed heatmaps}
\usage{
add_simply_logo(theme = "black")
}
\arguments{
\item{theme}{Options are "black" or "white"}
}
\description{
Add the logo in black or white depending on the theme
}

