% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_simplypar}
\alias{set_simplypar}
\title{Set the overall plotting parameters for the two simplystats themes}
\usage{
set_simplypar(theme = "black")
}
\arguments{
\item{theme}{Options are "black" or "white"}
}
\description{
The theme can be either black or white.
}

