% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplybox}
\alias{simplybox}
\title{A function for creating simplystats themed boxplots}
\usage{
simplybox(..., col = 1, logo = TRUE, theme = "black")
}
\arguments{
\item{...}{Pass any usual boxplot parameters}

\item{col}{A numeric vector specifying the colors of points}

\item{logo}{Logical, if true, then the simplystats logo is added to the plot}

\item{theme}{Options are "black" or "white"}
}
\description{
This function creates boxplots.  You can
also specify whether to add the simplystats logo or not.
}

