% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplyheat}
\alias{simplyheat}
\title{A function for creating simplystats themed heatmaps}
\usage{
simplyheat(dat, x = NULL, y = NULL, ..., ncol = 11, logo = TRUE,
  theme = "black")
}
\arguments{
\item{dat}{A data matrix}

\item{x}{The x axis labels, must be of length dim(dat)[2]}

\item{y}{The y axis labels, must be of length dim(dat)[1]}

\item{...}{Other arguments to the heatmap function}

\item{ncol}{The number of colors that you want}

\item{logo}{Logical, if true, then the simplystats logo is added to the plot}

\item{theme}{Options are "black" or "white"}
}
\description{
This function creates histograms. The default is a theme that runs
from blue (low) to orange (high). The heatmap is not clustered and
is oriented the "right" way so that the upper left cell of the matrix
is the upper left cell plotted in the heatmap.
}

