% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplyhist}
\alias{simplyhist}
\title{A function for creating simplystats themed histograms}
\usage{
simplyhist(..., col = 1, logo = TRUE, main = "", theme = "black")
}
\arguments{
\item{...}{Pass any usual histogram parameters}

\item{col}{A numeric vector specifying the colors of points}

\item{logo}{Logical, if true, then the simplystats logo is added to the plot}

\item{main}{The title of the plot, defaults to no title}

\item{theme}{Options are "black" or "white"}
}
\description{
This function creates histograms. The default histogram is simplystats blue
and there is no default title. You can
also specify whether to add the simplystats logo or not.
}

