% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplylines}
\alias{simplylines}
\title{A function for adding points using the simplystats theme}
\usage{
simplylines(..., col = 1)
}
\arguments{
\item{...}{Pass any usual plotting parameters to points}

\item{col}{A numeric vector specifying the colors of points}
}
\description{
This function creates scatterplots. To use it, you should pass a vector
of numeric color values (it will create a palette to match the length). You can
also specify whether to add the simplystats logo or not.
}

