% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplypoints}
\alias{simplypoints}
\title{A function for adding points using the simplystats theme}
\usage{
simplypoints(..., col = 1, pch = 19)
}
\arguments{
\item{...}{Pass any usual plotting parameters to points}

\item{col}{A numeric vector specifying the colors of points}

\item{pch}{The default pch is 19, cause it is the best.}
}
\description{
This function creates scatterplots. To use it, you should pass a vector
of numeric color values (it will create a palette to match the length). You can
also specify whether to add the simplystats logo or not.
}

