% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{simplystats}
\alias{simplystats}
\alias{simplystats-package}
\title{simplystats}
\description{
An R package for making simplystats themed plots
}
\details{
simplystats can be used to make plots in the style of the Simply Statistics
blog (http://simplystatistics.org,simplystats on twitter). The functions that are currently
available are \code{\link{simplyplot}} for scatterplots, along with\code{\link{simplylines}}
and \code{\link{simplypoints}} for adding lines and points. For histograms use
 \code{\link{simplyhist}}, for heatmaps \code{\link{simplyheat}}, and
for boxplots \code{\link{simplybox}}.
}
\author{
Jeffrey T. Leek
}
\references{
http://simplystatistics.org

http://twitter.com/simplystats
}

