% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simplyts}
\alias{simplyts}
\title{A function for creating simplystats themed time-series plots}
\usage{
simplyts(..., col = 1, pch = 19, logo = TRUE, theme = "black")
}
\arguments{
\item{...}{Pass any usual plotting parameters}

\item{col}{A numeric vector specifying the colors of points}

\item{pch}{The default pch is 19, cause it is the best.}

\item{logo}{Logical, if true, then the simplystats logo is added to the plot}

\item{theme}{Options are "black" or "white"}
}
\description{
This function creates time series plots. To use it, you should pass a vector
of numeric color values (it will create a palette to match the length). You can
also specify whether to add the simplystats logo or not.
}

